requirements.txt
Django==3.2.9
gunicorn==20.1.0
asgiref==3.4.1
whitenoise==5.3.0
sqlparse==0.4.2
requests==2.27.1
PyPDF2==1.27.9
gspread==5.3.2
docx-mailmerge==0.5.0
urllib3==1.26.9
python-docx==0.8.11
docxcompose==1.3.4
django-pwa==1.0.10
fcm-django==1.0.12
djangorestframework==3.13.1